backend/requirements.txt
# File Descriptor Forensics and Code Sandbox Backend
# Run: pip install -r backend/requirements.txt

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
google-generativeai>=0.3.0
reportlab>=4.0.0
